test: merge Vitest configs into unit and real-redis projects
