+++ package.json
@@ -60,11 +60,11 @@
   "scripts": {
     "build": "tsup src/index.ts src/cli.ts src/edge.ts --format esm,cjs --dts --clean",
     "build:all": "npm run build",
-    "test": "vitest run",
-    "test:coverage": "vitest run --coverage",
-    "test:integration": "vitest run --config vitest.integration.config.ts",
-    "test:all": "npm test && npm run test:integration",
-    "test:watch": "vitest",
+    "test": "vitest run --project unit",
+    "test:coverage": "vitest run --coverage --project unit",
+    "test:integration": "vitest run --project real-redis",
+    "test:all": "vitest run",
+    "test:watch": "vitest --project unit",
     "lint": "biome check .",
     "lint:fix": "biome check --write .",
     "bench:latency": "tsx benchmarks/latency.ts",
